feat: add DeepSeekService for another AI modal.
