+++ src/main/resources/application.properties
@@ -1,5 +1,11 @@
 spring.application.name=email-assistant-backend
 server.port=8081
 
+## Alibaba Gwen API meta data
 Gwen-fast.api.url=${GWEN_URL}
 Gwen-fast.api.key=${GWEN_KEY}
+
+## DeepSeek API meta data
+deepseek.api.url=${DEEPSEEK_URL}
+deepseek.api.key=${DEEPSEEK_KEY}
+
